gmsol_utils::market

Enum MarketConfigKey

#[non_exhaustive]

#[repr(u16)]
pub enum MarketConfigKey {

    SwapImpactExponent = 0,
    SwapImpactPositiveFactor = 1,
    SwapImpactNegativeFactor = 2,
    SwapFeeReceiverFactor = 3,
    SwapFeeFactorForPositiveImpact = 4,
    SwapFeeFactorForNegativeImpact = 5,
    MinPositionSizeUsd = 6,
    MinCollateralValue = 7,
    MinCollateralFactor = 8,
    MinCollateralFactorForOpenInterestMultiplierForLong = 9,
    MinCollateralFactorForOpenInterestMultiplierForShort = 10,
    MaxPositivePositionImpactFactor = 11,
    MaxNegativePositionImpactFactor = 12,
    MaxPositionImpactFactorForLiquidations = 13,
    PositionImpactExponent = 14,
    PositionImpactPositiveFactor = 15,
    PositionImpactNegativeFactor = 16,
    OrderFeeReceiverFactor = 17,
    OrderFeeFactorForPositiveImpact = 18,
    OrderFeeFactorForNegativeImpact = 19,
    LiquidationFeeReceiverFactor = 20,
    LiquidationFeeFactor = 21,
    PositionImpactDistributeFactor = 22,
    MinPositionImpactPoolAmount = 23,
    BorrowingFeeReceiverFactor = 24,
    BorrowingFeeFactorForLong = 25,
    BorrowingFeeFactorForShort = 26,
    BorrowingFeeExponentForLong = 27,
    BorrowingFeeExponentForShort = 28,
    BorrowingFeeOptimalUsageFactorForLong = 29,
    BorrowingFeeOptimalUsageFactorForShort = 30,
    BorrowingFeeBaseFactorForLong = 31,
    BorrowingFeeBaseFactorForShort = 32,
    BorrowingFeeAboveOptimalUsageFactorForLong = 33,
    BorrowingFeeAboveOptimalUsageFactorForShort = 34,
    FundingFeeExponent = 35,
    FundingFeeFactor = 36,
    FundingFeeMaxFactorPerSecond = 37,
    FundingFeeMinFactorPerSecond = 38,
    FundingFeeIncreaseFactorPerSecond = 39,
    FundingFeeDecreaseFactorPerSecond = 40,
    FundingFeeThresholdForStableFunding = 41,
    FundingFeeThresholdForDecreaseFunding = 42,
    ReserveFactor = 43,
    OpenInterestReserveFactor = 44,
    MaxPnlFactorForLongDeposit = 45,
    MaxPnlFactorForShortDeposit = 46,
    MaxPnlFactorForLongWithdrawal = 47,
    MaxPnlFactorForShortWithdrawal = 48,
    MaxPnlFactorForLongTrader = 49,
    MaxPnlFactorForShortTrader = 50,
    MaxPnlFactorForLongAdl = 51,
    MaxPnlFactorForShortAdl = 52,
    MinPnlFactorAfterLongAdl = 53,
    MinPnlFactorAfterShortAdl = 54,
    MaxPoolAmountForLongToken = 55,
    MaxPoolAmountForShortToken = 56,
    MaxPoolValueForDepositForLongToken = 57,
    MaxPoolValueForDepositForShortToken = 58,
    MaxOpenInterestForLong = 59,
    MaxOpenInterestForShort = 60,
    MinTokensForFirstDeposit = 61,

}

Market config keys.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

SwapImpactExponent = 0

Swap impact exponent.

SwapImpactPositiveFactor = 1

Swap impact positive factor.

SwapImpactNegativeFactor = 2

Swap impact negative factor.

SwapFeeReceiverFactor = 3

Swap fee receiver factor.

SwapFeeFactorForPositiveImpact = 4

Swap fee factor for positive impact.

SwapFeeFactorForNegativeImpact = 5

Swap fee factor for negative impact.

MinPositionSizeUsd = 6

Min position size usd.

MinCollateralValue = 7

Min collateral value.

MinCollateralFactor = 8

Min collateral factor.

MinCollateralFactorForOpenInterestMultiplierForLong = 9

Min collateral factor for open interest multiplier for long.

MinCollateralFactorForOpenInterestMultiplierForShort = 10

Min collateral factor for open interest multiplier for short.

MaxPositivePositionImpactFactor = 11

Max positive position impact factor.

MaxNegativePositionImpactFactor = 12

Max negative position impact factor.

MaxPositionImpactFactorForLiquidations = 13

Max position impact factor for liquidations.

PositionImpactExponent = 14

Position impact exponent.

PositionImpactPositiveFactor = 15

Position impact positive factor.

PositionImpactNegativeFactor = 16

Position impact negative factor.

OrderFeeReceiverFactor = 17

Order fee receiver factor.

OrderFeeFactorForPositiveImpact = 18

Order fee factor for positive impact.

OrderFeeFactorForNegativeImpact = 19

Order fee factor for negative impact.

LiquidationFeeReceiverFactor = 20

Liquidation fee receiver factor.

LiquidationFeeFactor = 21

Liquidation fee factor.

PositionImpactDistributeFactor = 22

Position impact distribute factor.

MinPositionImpactPoolAmount = 23

Min position impact pool amount.

BorrowingFeeReceiverFactor = 24

Borrowing fee receiver factor.

BorrowingFeeFactorForLong = 25

Borrowing fee factor for long.

BorrowingFeeFactorForShort = 26

Borrowing fee factor for short.

BorrowingFeeExponentForLong = 27

Borrowing fee exponent for long.

BorrowingFeeExponentForShort = 28

Borrowing fee exponent for short.

BorrowingFeeOptimalUsageFactorForLong = 29

Borrowing fee optimal usage factor for long.

BorrowingFeeOptimalUsageFactorForShort = 30

Borrowing fee optimal usage factor for short.

BorrowingFeeBaseFactorForLong = 31

Borrowing fee base factor for long.

BorrowingFeeBaseFactorForShort = 32

Borrowing fee base factor for short.

BorrowingFeeAboveOptimalUsageFactorForLong = 33

Borrowing fee above optimal usage factor for long.

BorrowingFeeAboveOptimalUsageFactorForShort = 34

Borrowing fee above optimal usage factor for short.

FundingFeeExponent = 35

Funding fee exponent.

FundingFeeFactor = 36

Funding fee factor.

FundingFeeMaxFactorPerSecond = 37

Funding fee max factor per second.

FundingFeeMinFactorPerSecond = 38

Funding fee min factor per second.

FundingFeeIncreaseFactorPerSecond = 39

Funding fee increase factor per second.

FundingFeeDecreaseFactorPerSecond = 40

Funding fee decrease factor per second.

FundingFeeThresholdForStableFunding = 41

Funding fee threshold for stable funding.

FundingFeeThresholdForDecreaseFunding = 42

Funding fee threshold for decrease funding.

ReserveFactor = 43

Reserve factor.

OpenInterestReserveFactor = 44

Open interest reserve factor.

MaxPnlFactorForLongDeposit = 45

Max PNL factor for long deposit.

MaxPnlFactorForShortDeposit = 46

Max PNL factor for short deposit.

MaxPnlFactorForLongWithdrawal = 47

Max PNL factor for long withdrawal.

MaxPnlFactorForShortWithdrawal = 48

Max PNL factor for short withdrawal.

MaxPnlFactorForLongTrader = 49

Max PNL factor for long trader.

MaxPnlFactorForShortTrader = 50

Max PNL factor for short trader.

MaxPnlFactorForLongAdl = 51

Max PNL factor for long ADL.

MaxPnlFactorForShortAdl = 52

Max PNL factor for short ADL.

MinPnlFactorAfterLongAdl = 53

Min PNL factor after long ADL.

MinPnlFactorAfterShortAdl = 54

Min PNL factor after short ADL.

MaxPoolAmountForLongToken = 55

Max pool amount for long token.

MaxPoolAmountForShortToken = 56

Max pool amount for short token.

MaxPoolValueForDepositForLongToken = 57

Max pool value for deposit for long token.

MaxPoolValueForDepositForShortToken = 58

Max pool value for deposit for short token.

MaxOpenInterestForLong = 59

Max open interest for long.

MaxOpenInterestForShort = 60

Max open interest for short.

MinTokensForFirstDeposit = 61

Min tokens for first deposit.